+++ src/main/java/shadows/apotheosis/adventure/affix/socket/gem/GemInstance.java
@@ -3,25 +3,17 @@ package shadows.apotheosis.adventure.affix.socket.gem;
 import net.minecraft.world.item.ItemStack;
 import shadows.apotheosis.adventure.loot.LootRarity;
 
-public record GemInstance(ItemStack gemStack, Gem gem, LootRarity rarity, int facets) {
+public record GemInstance(ItemStack gemStack, Gem gem, LootRarity rarity) {
 
 	public GemInstance(ItemStack gemStack) {
-		this(gemStack, GemItem.getGem(gemStack), GemItem.getLootRarity(gemStack), GemItem.getFacets(gemStack));
+		this(gemStack, GemItem.getGem(gemStack), GemItem.getLootRarity(gemStack));
 	}
 
 	public boolean isValid() {
-		return this.gem != null && this.rarity != null && facets >= 0 && facets <= this.gem.getMaxFacets(rarity);
+		return this.gem != null && this.rarity != null;
 	}
 
 	public boolean isValidIn(ItemStack socketed) {
 		return isValid() && this.gem.isValidIn(socketed, gemStack, rarity);
-	}
-
-	public int maxFacets() {
-		return this.gem.getMaxFacets(this.rarity);
-	}
-
-	public boolean isMaxed() {
-		return this.facets == this.maxFacets();
 	}
 }